requirements.txt
einops==0.7.0
matplotlib==3.8.2
open-clip-torch==2.24.0
PyYAML=6.0.1
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
tensorboard-data-server==0.7.2
timm==0.9.12
tokenizers==0.15.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.37.2